+++ client/index.html
@@ -6,6 +6,10 @@
     <title>Document</title>
 </head>
 <body style="height: 100vh;">
+    <ul></ul>
+    <input placeholder="message">
+    <button>Send</button>
+    <script src="https://cdn.socket.io/socket.io-3.0.0.js"></script>
     <script src="index.js"></script>
 </body>
 </html>
