+++ server/package.json
@@ -10,6 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "socket.io": "^4.7.2",
     "ws": "^8.14.1"
   }
 }
